requirements.txt
click
pillow
requests
tqdm